Refactor out final pieces and use `_with` pragma to keep consistency
